LayoutTests/http/tests/resourceLoadStatistics/non-prevalent-resources-can-access-cookies-in-a-third-party-context-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
  

--------
Frame: '<!--frame1-->'
--------
Should receive one cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=value

--------
Frame: '<!--frame2-->'
--------
Setting partitioned, third party cookie.


--------
Frame: '<!--frame3-->'
--------
Should receive one partitioned, third party cookie.
Did not receive cookie named 'firstPartyCookie'.
Received cookie named 'partitionedCookie'.
Client-side document.cookie: partitionedCookie=value

--------
Frame: '<!--frame4-->'
--------
After user interaction, should receive one non-partitioned, first party cookie.
Received cookie named 'firstPartyCookie'.
Did not receive cookie named 'partitionedCookie'.
Client-side document.cookie: firstPartyCookie=value
